Updated minigame
First try at shooting and scoring - enemies have three health and projectiles are limited by time not quantity. Note: The update method is new and used to update the state of enemies and projectiles - it will be tweaked in the next update but currently returns true when the sprite is to be destroyed. Note: must .remove sprite from lists, .kill doesn't work for some reason