The sweet sweet release of death
Players now have health that decreases when enemies impact them or go
offscreen. When it hits zero, there’s a game-over “menu’ that appears,
prompting them to restart. There is a not-entirely-functional and
currently commented out new way to make enemies move towards the player
when they spawn (This probably requires trig to do properly)